requirements.txt
# torch==1.13.1
# torchvision==0.14.1
pytorch-lightning==1.9
lightning-bolts
mitsuba==3.5.1

opencv-python==4.7.0.72
Pillow==9.5.0
trimesh==3.21.4
matplotlib==3.7.1
scipy==1.10.1
tensorboardX
jupyter
ipywidgets
open3d
omegaconf
scikit-learn
nerfstudio
lightning-bolts
torchviz
openexr_numpy
